fix: clear qodana warning annotations
