+++ src/relay_teams/agents/orchestration/task_execution_service.py
@@ -6,7 +6,7 @@ import json
 import logging
 from collections.abc import Callable, Mapping
 from pathlib import Path
-from typing import TYPE_CHECKING, Literal, cast
+from typing import Literal, cast
 
 from pydantic import BaseModel, ConfigDict, JsonValue
 from pydantic_ai.messages import ModelRequest, UserContent, UserPromptPart
@@ -65,34 +65,23 @@ from relay_teams.sessions.runs.assistant_errors import (
     build_assistant_error_message,
     build_assistant_error_response,
 )
+from relay_teams.agents.orchestration.task_contracts import TaskExecutionResult
 from relay_teams.sessions.runs.run_runtime_repo import (
     RunRuntimePhase,
     RunRuntimeRepository,
     RunRuntimeStatus,
 )
 from relay_teams.hooks import HookEventName, HookService, TaskCompletedInput
-from relay_teams.tools.registry.registry import ToolResolutionContext
-
-if TYPE_CHECKING:
-    from relay_teams.skills.skill_registry import SkillRegistry
-    from relay_teams.skills.skill_models import SkillInstructionEntry
-    from relay_teams.skills.skill_routing_service import SkillRuntimeService
-    from relay_teams.tools.registry import ToolRegistry
+from relay_teams.skills.skill_models import SkillInstructionEntry
+from relay_teams.skills.skill_registry import SkillRegistry
+from relay_teams.skills.skill_routing_service import SkillRuntimeService
+from relay_teams.tools.registry.registry import ToolRegistry, ToolResolutionContext
 from relay_teams.tools.runtime.approval_ticket_repo import ApprovalTicketRepository
 from relay_teams.agents.instances.instance_repository import AgentInstanceRepository
 from relay_teams.workspace import WorkspaceHandle, WorkspaceManager
 
 LOGGER = get_logger(__name__)
 ProviderUserPromptContent = str | tuple[UserContent, ...]
-
-
-class TaskExecutionResult(BaseModel):
-    model_config = ConfigDict(extra="forbid", frozen=True)
-
-    output: str
-    completion_reason: RunCompletionReason = RunCompletionReason.ASSISTANT_RESPONSE
-    error_code: str | None = None
-    error_message: str | None = None
 
 
 class TaskExecutionService(BaseModel):
@@ -740,8 +729,8 @@ class TaskExecutionService(BaseModel):
         role: RoleDefinition,
         task: TaskEnvelope | None = None,
     ) -> RuntimeToolsSnapshot:
-        skill_registry = cast("SkillRegistry", self.skill_registry)
-        tool_registry = cast("ToolRegistry", self.tool_registry)
+        skill_registry = cast(SkillRegistry, self.skill_registry)
+        tool_registry = cast(ToolRegistry, self.tool_registry)
         resolved_skills = skill_registry.resolve_known(
             role.skills,
             strict=False,
@@ -993,7 +982,7 @@ class TaskExecutionService(BaseModel):
                 (),
             )
         skill_runtime_service = cast(
-            "SkillRuntimeService",
+            SkillRuntimeService,
             self.skill_runtime_service,
         )
         prepared_prompt = skill_runtime_service.prepare_prompt(
@@ -1013,7 +1002,7 @@ class TaskExecutionService(BaseModel):
 
     def _to_prompt_skill_instructions(
         self,
-        entries: tuple["SkillInstructionEntry", ...],
+        entries: tuple[SkillInstructionEntry, ...],
     ) -> tuple[PromptSkillInstruction, ...]:
         return tuple(
             PromptSkillInstruction(name=entry.name, description=entry.description)
