+++ src/relay_teams/automation/automation_bound_session_queue_service.py
@@ -76,7 +76,8 @@ class FeishuRuntimeConfigLookup(Protocol):
 
 class FeishuRuntimeConfigLike(Protocol):
     @property
-    def environment(self) -> FeishuEnvironment: ...
+    def environment(self) -> FeishuEnvironment:
+        raise NotImplementedError
 
 
 class FeishuClientLike(Protocol):
