+++ src/relay_teams/automation/automation_delivery_service.py
@@ -48,7 +48,8 @@ class FeishuRuntimeConfigLookup(Protocol):
 
 class FeishuRuntimeConfigLike(Protocol):
     @property
-    def environment(self) -> FeishuEnvironment: ...
+    def environment(self) -> FeishuEnvironment:
+        raise NotImplementedError
 
 
 class FeishuClientLike(Protocol):
